--- a/sdk/interpolate.go
+++ b/sdk/interpolate.go
@@ -12,9 +12,13 @@ var interpolateRegex = regexp.MustCompile("({{\\.[a-zA-Z0-9._|\\s]+}})")
 
 // Interpolate returns interpolated input with vars
 func Interpolate(input string, vars map[string]string) (string, error) {
-	data := map[string]string{}
-	defaults := map[string]string{}
-	empty := map[string]string{}
+	if !strings.Contains(input, "{{") {
+		return input, nil
+	}
+
+	data := make(map[string]string, len(vars))
+	defaults := make(map[string]string, len(vars))
+	empty := make(map[string]string, len(vars))
 
 	for k, v := range vars {
 		kb := strings.Replace(k, ".", "__", -1)