before
````
Running tool: /Applications/go/bin/go test -v -benchmem -run=^$ github.com/ovh/cds/sdk -bench ^BenchmarkInterpolate$

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk
BenchmarkInterpolate-8   	    5000	    261920 ns/op	  109879 B/op	     682 allocs/op
PASS
ok  	github.com/ovh/cds/sdk	1.352s
Success: Benchmarks passed.
````

after 
````
Running tool: /Applications/go/bin/go test -v -benchmem -run=^$ github.com/ovh/cds/sdk -bench ^BenchmarkInterpolate$

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk
BenchmarkInterpolate-8   	   10000	    152740 ns/op	   59311 B/op	     560 allocs/op
PASS
ok  	github.com/ovh/cds/sdk	1.558s
Success: Benchmarks passed.
````